refactor canvas and extract drawing utilities
